move getComponents into gcloud
