+++ src/main/java/com/google/cloud/tools/appengine/cloudsdk/CloudSdk.java
@@ -16,15 +16,10 @@
 
 package com.google.cloud.tools.appengine.cloudsdk;
 
-import com.google.cloud.tools.appengine.cloudsdk.internal.args.GcloudArgs;
-import com.google.cloud.tools.appengine.cloudsdk.process.ProcessHandlerException;
-import com.google.cloud.tools.appengine.cloudsdk.serialization.CloudSdkComponent;
 import com.google.cloud.tools.appengine.cloudsdk.serialization.CloudSdkVersion;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Joiner;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
-import com.google.gson.JsonSyntaxException;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
@@ -103,34 +98,6 @@ public class CloudSdk {
       throw new CloudSdkVersionFileParseException(
           "Pattern found in the Cloud SDK version file could not be parsed: " + contents, ex);
     }
-  }
-
-  /**
-   * Returns the list of Cloud SDK Components and their settings, reported by the current gcloud
-   * installation. Unlike other methods in this class that call gcloud, this method always uses a
-   * synchronous ProcessRunner and will block until the gcloud process returns.
-   *
-   * @throws ProcessHandlerException when process runner encounters an error
-   * @throws JsonSyntaxException when the cloud SDK output cannot be parsed
-   * @throws CloudSdkNotFoundException when the Cloud SDK is not installed where expected
-   * @throws CloudSdkOutOfDateException when the installed Cloud SDK is too old
-   */
-  public List<CloudSdkComponent> getComponents()
-      throws ProcessHandlerException, JsonSyntaxException, CloudSdkNotFoundException,
-          CloudSdkOutOfDateException, CloudSdkVersionFileException, InvalidJavaSdkException,
-          IOException {
-    validateCloudSdk();
-
-    // gcloud components list --show-versions --format=json
-    List<String> command =
-        new ImmutableList.Builder<String>()
-            .add("components", "list")
-            .addAll(GcloudArgs.get("show-versions", true))
-            .addAll(GcloudArgs.get("format", "json"))
-            .build();
-
-    String componentsJson = Gcloud.builder(this).build().runCommand(command);
-    return CloudSdkComponent.fromJsonList(componentsJson);
   }
 
   public Path getSdkPath() {
